add framework for venue setting
The recent venues dataset I think is built nightly, so isn't going to
work quite yet on new venues. However the bulk of this code should be
functioning now. It's got a hardcoded slug at the moment, because I'm
not yet ready to add another setting. In the drupal version of this we
actually set the slug in settings and did a round trip to save the id
during settings update.